Explores some 1d self-gravity ideas
I realized that the FFT self-gravity thing really doesn't make sense
when there's a net charge, which of course there is in gravity. There's
a class in gravity.py which adds 1/2 rhobar x^2 in order to effectively
build a sine series solution. It has not been tested yet in a dynamic
gravity problem.

The larger issue here is that the gravp and grave systems assume that
the Laplacian of the potential REALLY is the density at every
point. That assumption is not there in source-terms gravity (gravs
system).