Demonstrates a good, working 2d driving field based on FFT's
Note that this driving field might be for demonstration only, as it's
probably not useful for parallel jobs. That's because there's really no
good way to generate random numbers uniformly across all cores in a
consistent way. Options are to use (and serialize) a single RNG stream
per wave-number which seems very wasteful, or to have all processes
execute the same global RNG sequence, while only extracting the deviates
corresponding to the Fourier modes they own.