Implement coarse-to-fine guard transfer for a special case
8 coarse blocks are created, one of which (block 3) has two child
blocks. The solution can be evolved passively on the finer grids, in
other words it is never projected back to the parent block. The C code
assumes a very specialized case: 1d, 3 guard zones, and child blocks
which cover the parent block (i.e. the starting index into the parent
block can be ignored).